backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.27.2
apscheduler==3.10.4
python-multipart==0.0.9
playwright==1.49.1
